requirements.txt
amqp==5.2.0
asgiref==3.7.2
astroid==2.15.8
billiard==4.2.0
black==24.1.1
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.4.1
Django==5.0.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
flake8==6.1.0
flake8-django==1.4
kombu==5.3.5
lazy-object-proxy==1.10.0
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pathspec==0.12.1
platformdirs==4.2.0
prompt-toolkit==3.0.43
pycodestyle==2.11.1
pyflakes==3.1.0
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
